src/core/coreml_wrapper.py
"""
Core ML / Neural Engine wrapper for Apple Silicon acceleration.
Uses Apple's Neural Engine for additional ML acceleration.
"""

from typing import Optional, List, Dict
from loguru import logger
import numpy as np

# Try to import Core ML
try:
    import coremltools as ct
    COREML_AVAILABLE = True
except ImportError:
    COREML_AVAILABLE = False
    logger.debug("coremltools not available")

# Try to import Neural Engine utilities
try:
    from coremltools.models.neural_network import quantization_utils
    NEURAL_ENGINE_AVAILABLE = True
except ImportError:
    NEURAL_ENGINE_AVAILABLE = False


class CoreMLWrapper:
    """
    Wrapper for Core ML models to use Neural Engine.
    Provides fallback to MPS/CPU if Core ML not available.
    """
    
    def __init__(self, model_name: str, use_neural_engine: bool = True):
        """
        Initialize Core ML wrapper.
        
        Args:
            model_name: Name of the model to load
            use_neural_engine: Whether to use Neural Engine (if available)
        """
        self.model_name = model_name
        self.use_neural_engine = use_neural_engine
        self.coreml_model = None
        self.fallback_model = None
        self.device = "neural_engine" if use_neural_engine else "mps"
        
        if COREML_AVAILABLE and use_neural_engine:
            self._load_coreml_model()
        else:
            self._load_fallback_model()
    
    def _load_coreml_model(self):
        """Load Core ML model for Neural Engine."""
        try:
            # Try to load pre-converted Core ML model
            # Note: Models need to be converted to Core ML format first
            # This is a placeholder - actual conversion requires model-specific code
            logger.info(f"Core ML support available but model conversion needed for {self.model_name}")
            logger.info("Falling back to MPS/CPU. To use Neural Engine, convert models to Core ML format.")
            self._load_fallback_model()
        except Exception as e:
            logger.warning(f"Failed to load Core ML model: {e}. Using fallback.")
            self._load_fallback_model()
    
    def _load_fallback_model(self):
        """Load fallback model (MPS or CPU)."""
        try:
            from sentence_transformers import SentenceTransformer
            # Use MPS if available, otherwise CPU
            import torch
            if torch.backends.mps.is_available():
                device = "mps"
            else:
                device = "cpu"
            
            self.fallback_model = SentenceTransformer(self.model_name, device=device)
            self.device = device
            logger.info(f"Loaded fallback model {self.model_name} on {device}")
        except Exception as e:
            logger.error(f"Failed to load fallback model: {e}")
            raise
    
    def encode(self, texts: List[str], batch_size: int = 32, **kwargs) -> np.ndarray:
        """
        Encode texts to embeddings.
        
        Args:
            texts: List of text strings
            batch_size: Batch size for processing
            **kwargs: Additional arguments
            
        Returns:
            Numpy array of embeddings
        """
        if self.coreml_model:
            # Use Core ML model (if available)
            return self._encode_coreml(texts, batch_size, **kwargs)
        else:
            # Use fallback model
            return self.fallback_model.encode(
                texts,
                batch_size=batch_size,
                show_progress_bar=False,
                **kwargs
            )
    
    def _encode_coreml(self, texts: List[str], batch_size: int, **kwargs) -> np.ndarray:
        """Encode using Core ML model (placeholder - needs implementation)."""
        # This would use the Core ML model for inference
        # Requires model-specific implementation
        logger.warning("Core ML encoding not yet implemented, using fallback")
        return self.fallback_model.encode(texts, batch_size=batch_size, **kwargs)
    
    @staticmethod
    def convert_model_to_coreml(model_name: str, output_path: str) -> bool:
        """
        Convert a sentence transformer model to Core ML format.
        
        Args:
            model_name: Name of the model to convert
            output_path: Path to save Core ML model
            
        Returns:
            True if successful, False otherwise
        """
        if not COREML_AVAILABLE:
            logger.error("coremltools not available, cannot convert model")
            return False
        
        try:
            from sentence_transformers import SentenceTransformer
            
            logger.info(f"Converting {model_name} to Core ML format...")
            model = SentenceTransformer(model_name)
            
            # Convert to Core ML (this is a simplified example)
            # Actual conversion requires more complex handling
            logger.warning("Core ML conversion requires model-specific implementation")
            logger.info("For now, using MPS (Metal) which provides excellent performance on Apple Silicon")
            
            return False  # Not fully implemented yet
        except Exception as e:
            logger.error(f"Failed to convert model: {e}")
            return False


def is_neural_engine_available() -> bool:
    """Check if Neural Engine is available."""
    return COREML_AVAILABLE and NEURAL_ENGINE_AVAILABLE


def get_optimal_device() -> str:
    """
    Get optimal device for ML operations on Apple Silicon.
    
    Returns:
        Device string: 'neural_engine', 'mps', or 'cpu'
    """
    if is_neural_engine_available():
        return "neural_engine"
    
    try:
        import torch
        if torch.backends.mps.is_available():
            return "mps"
    except ImportError:
        pass
    
    return "cpu"

